requirement.txt
Flask==3.0.3
Flask_Cors==4.0.1
pypdf==4.3.1
tiktoken==0.7.0
pandas==2.2.2
langchain==0.2.10
langchain-community==0.2.9
atlassian-python-api==3.41.14
huggingface-hub==0.24.0
torch==2.2.0
langchain-groq==0.1.6
groq==0.9.0
langchain_core==0.2.22
mysql-connector-python==9.0.0
python-dotenv
